+++ CHANGELOG.md
@@ -33,6 +33,10 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
   [Samuel Ford](https://github.com/samuelwford)
   [#1042](https://github.com/CocoaPods/CocoaPods/issues/1042)
 
+* `pod update` prints the previous version of the updated pods  
+  [Andrea Mazzini](https://github.com/andreamazz)
+  [#2008](https://github.com/CocoaPods/CocoaPods/issues/2008)
+
 ## 0.31.1
 [CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.31.1...0.31.0)
 • [CocoaPods-Core](https://github.com/CocoaPods/Core/compare/0.31.1...0.31.0)